Shapefile : Fix reading of huge dbf files (over 2GB) using long value in index position

+++ modules/datastore/geotk-dbf/src/main/java/org/geotoolkit/data/dbf/DbaseFileReader.java
@@ -117,7 +117,7 @@ public class DbaseFileReader {
 
     protected boolean randomAccessEnabled;
 
-    protected int currentOffset = 0;
+    protected long currentOffset = 0;
 
     private Charset stringCharset;
